src/requirements.txt
Flask-SocketIO==4.3.1
Flask==2.0.3
python-engineio==3.13.2
python-socketio==4.6.0
Werkzeug==2.0.3
opencv_python==4.7.0.68
numpy==1.24.2
ultralytics
PyJWT
python-dotenv
gunicorn
eventlet